+++ EssentialFeed/EssentialFeediOSTests/FeedSnapshotTests.swift
@@ -92,175 +92,4 @@ final class FeedSnapshotTests: XCTestCase {
             )
         ]
     }
-    
-    private func record(
-        snapshot: UIImage,
-        named: String,
-        file: StaticString = #file,
-        line: UInt = #line
-    ) {
-        let snapshotData = makeSnapshotData(image: snapshot, file: file, line: line)
-        let snapshotURL = makeSnapshotURL(named: named, file: file, line: line)
-        
-        do {
-            try FileManager.default.createDirectory(
-                at: snapshotURL.deletingLastPathComponent(),
-                withIntermediateDirectories: true
-            )
-            
-            try snapshotData?.write(to: snapshotURL)
-        } catch {
-            XCTFail("Failed to record snapshot with error: \(error)", file: file, line: line)
-        }
-    }
-    
-    private func assert(
-        snapshot: UIImage,
-        named: String,
-        file: StaticString = #file,
-        line: UInt = #line
-    ) {
-        let snapshotData = makeSnapshotData(image: snapshot)
-        let snapshotURL = makeSnapshotURL(named: named)
-        let storedSnapshotData = loadSnapshot(from: snapshotURL)
-        
-        if snapshotData != storedSnapshotData {
-            let tempPath = URL(filePath: NSTemporaryDirectory())
-                .appendingPathComponent(snapshotURL.lastPathComponent + ".png")
-            
-            try? snapshotData?.write(to: tempPath)
-            
-            XCTFail("New snapshot does not match stored snapshot, new snapshot saved at \(tempPath)", file: file, line: line)
-        }
-    }
-    
-    private func makeSnapshotURL(named: String, file: StaticString = #file, line: UInt = #line) -> URL {
-        let snapshotURL = URL(filePath: String(describing: file))
-            .deletingLastPathComponent()
-            .appendingPathComponent("snapshots")
-            .appendingPathComponent(named + ".png")
-        
-        return snapshotURL
-    }
-    
-    private func makeSnapshotData(image: UIImage, file: StaticString = #file, line: UInt = #line) -> Data? {
-        guard let pngRepresentation = image.pngData() else {
-            XCTFail("Failed to convert image to PNG data", file: file, line: line)
-            return nil
-        }
-        
-        return pngRepresentation
-    }
-    
-    private func loadSnapshot(from url: URL, file: StaticString = #file, line: UInt = #line) -> Data? {
-        guard let storedSnapshotData = try? Data(NSData(contentsOf: url)) else {
-            XCTFail("Failed to load snapshot at \(url). Use `record` to record first", file: file, line: line)
-            return nil
-        }
-        
-        return storedSnapshotData
-    }
 }
-
-struct SnapshotConfiguration {
-    let size: CGSize
-    let safeAreaInsets: UIEdgeInsets
-    let layoutMargins: UIEdgeInsets
-    let traitCollection: UITraitCollection
-    
-    static func iphone8(style: UIUserInterfaceStyle) -> SnapshotConfiguration {
-        SnapshotConfiguration(
-            size: CGSize(width: 375, height: 667),
-            safeAreaInsets: UIEdgeInsets(top: 20, left: 0, bottom: 0, right: 0),
-            layoutMargins: UIEdgeInsets(top: 20, left: 16, bottom: 0, right: 16),
-            traitCollection: UITraitCollection(mutations: { mutableTraits in
-                mutableTraits.forceTouchCapability = .available
-                mutableTraits.layoutDirection = .leftToRight
-                mutableTraits.userInterfaceIdiom = .phone
-                mutableTraits.horizontalSizeClass = .compact
-                mutableTraits.verticalSizeClass = .regular
-                mutableTraits.displayScale = 2
-                mutableTraits.displayGamut = .P3
-                mutableTraits.userInterfaceStyle = style
-            })
-        )
-    }
-}
-
-private final class SnapshotWindow: UIWindow {
-    private var configuration: SnapshotConfiguration = .iphone8(style: .light)
-    
-    convenience init(configuration: SnapshotConfiguration, root: UIViewController) {
-        self.init(frame: CGRect(origin: .zero, size: configuration.size))
-        
-        self.configuration = configuration
-        self.layoutMargins = configuration.layoutMargins
-        self.rootViewController = root
-        self.isHidden = false
-        root.view.layoutMargins = configuration.layoutMargins
-    }
-    
-    override var safeAreaInsets: UIEdgeInsets {
-        configuration.safeAreaInsets
-    }
-    
-    override var traitCollection: UITraitCollection {
-        UITraitCollection { mutableTraits in
-            mutableTraits.forceTouchCapability = configuration.traitCollection.forceTouchCapability
-            mutableTraits.layoutDirection = configuration.traitCollection.layoutDirection
-            mutableTraits.userInterfaceIdiom = configuration.traitCollection.userInterfaceIdiom
-            mutableTraits.horizontalSizeClass = configuration.traitCollection.horizontalSizeClass
-            mutableTraits.verticalSizeClass = configuration.traitCollection.verticalSizeClass
-            mutableTraits.displayScale = configuration.traitCollection.displayScale
-            mutableTraits.displayGamut = configuration.traitCollection.displayGamut
-            mutableTraits.userInterfaceStyle = configuration.traitCollection.userInterfaceStyle
-        }
-    }
-    
-    func snapshot() -> UIImage {
-        let renderer = UIGraphicsImageRenderer(bounds: bounds, format: .init(for: traitCollection))
-        return renderer.image { action in
-            layer.render(in: action.cgContext)
-        }
-    }
-}
-
-private class ImageStub: FeedImageCellControllerDelegate {
-    weak var controller: FeedImageCellController?
-    private let viewModel: FeedImageViewModel<UIImage>
-    
-    init(description: String?, location: String?, image: UIImage?) {
-        self.viewModel = FeedImageViewModel(
-            location: location,
-            description: description,
-            image: image,
-            isLoading: false,
-            shouldRetry: image == nil
-        )
-    }
-    
-    func didRequestImage() {
-        controller?.display(viewModel: viewModel)
-    }
-    
-    func didCancelImageRequest() {}
-}
-
-private extension FeedViewController {
-    func display(_ stubs: [ImageStub]) {
-        self.display(
-            stubs.map { stub in
-                let controller = FeedImageCellController(delegate: stub)
-                stub.controller = controller
-                return controller
-            }
-        )
-    }
-}
-
-extension UIViewController {
-    func snapshot(for configuration: SnapshotConfiguration) -> UIImage {
-        SnapshotWindow(configuration: configuration, root: self).snapshot()
-    }
-}
-
