+++ EssentialFeed/EssentialFeediOSTests/Helpers/SnapshotWindow.swift
@@ -1,0 +1,45 @@
+//
+//  SnapshotWindow.swift
+//  EssentialFeed
+//
+//  Created by Anthony on 26/7/25.
+//
+import UIKit
+
+final class SnapshotWindow: UIWindow {
+    private var configuration: SnapshotConfiguration = .iphone8(style: .light)
+    
+    convenience init(configuration: SnapshotConfiguration, root: UIViewController) {
+        self.init(frame: CGRect(origin: .zero, size: configuration.size))
+        
+        self.configuration = configuration
+        self.layoutMargins = configuration.layoutMargins
+        self.rootViewController = root
+        self.isHidden = false
+        root.view.layoutMargins = configuration.layoutMargins
+    }
+    
+    override var safeAreaInsets: UIEdgeInsets {
+        configuration.safeAreaInsets
+    }
+    
+    override var traitCollection: UITraitCollection {
+        UITraitCollection { mutableTraits in
+            mutableTraits.forceTouchCapability = configuration.traitCollection.forceTouchCapability
+            mutableTraits.layoutDirection = configuration.traitCollection.layoutDirection
+            mutableTraits.userInterfaceIdiom = configuration.traitCollection.userInterfaceIdiom
+            mutableTraits.horizontalSizeClass = configuration.traitCollection.horizontalSizeClass
+            mutableTraits.verticalSizeClass = configuration.traitCollection.verticalSizeClass
+            mutableTraits.displayScale = configuration.traitCollection.displayScale
+            mutableTraits.displayGamut = configuration.traitCollection.displayGamut
+            mutableTraits.userInterfaceStyle = configuration.traitCollection.userInterfaceStyle
+        }
+    }
+    
+    func snapshot() -> UIImage {
+        let renderer = UIGraphicsImageRenderer(bounds: bounds, format: .init(for: traitCollection))
+        return renderer.image { action in
+            layer.render(in: action.cgContext)
+        }
+    }
+}
